split vuex store
